`c2rust-analyze`: Move all of `main.rs` to `analyze.rs` except for `fn main` to prepare for putting the `cargo` wrapper in `main.rs`.
